+++ app/src/main/java/ru/kostenko/nework/viewmodel/UserViewModel.kt
@@ -33,7 +33,7 @@ class UserViewModel @Inject constructor(
     val dataSetPeople: LiveData<List<User>> =
         userRepositoryImpl.data.map { list ->
             list.mapNotNull { user ->
-                if(userIds.value?.contains(user.id)!!) user else null
+                if(userIds.value?.contains(user.id.toLong())!!) user else null
             }
         }.asLiveData(Dispatchers.Default)
 
@@ -45,8 +45,8 @@ class UserViewModel @Inject constructor(
     val user: LiveData<User>
         get() = _user
 
-    private val _userIds = MutableLiveData<Set<Int>>()
-    val userIds: LiveData<Set<Int>>
+    private val _userIds = MutableLiveData<Set<Long>>()
+    val userIds: LiveData<Set<Long>>
         get() = _userIds
 
     init {
@@ -73,7 +73,7 @@ class UserViewModel @Inject constructor(
         }
     }
 
-    fun getUsersIds(set: Set<Int>) =
+    fun getUsersIds(set: Set<Long>) =
         viewModelScope.launch { _userIds.value = set }
 
     fun getTakeble(): LiveData<List<User>> =
@@ -83,7 +83,7 @@ class UserViewModel @Inject constructor(
             }
         }.asLiveData(Dispatchers.Default)
 
-    fun setSetIds(set: Set<Int>) {
+    fun setSetIds(set: Set<Long>) {
         _userIds.value = set
     }
